requirements.txt
ctransformers==0.2.5
# CTransformers
sentence-transformers==2.2.2
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain-pinecone
langchain_community
-e .